% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confcurve.R
\name{bootcurve}
\alias{bootcurve}
\title{Generate bootstrap samples for computation of confidence curves and related inferential statistics.}
\usage{
bootcurve(data, statistic, B = 2000, formula = NULL)
}
\arguments{
\item{data}{the data frame containing the original data set.}

\item{statistic}{a function specifying the sample statistic to be computed, in the format expected by the boot package.}

\item{B}{the number of bootstrap samples to generate.}

\item{formula}{a formula, if needed by the statistic function}
}
\description{
A function to generate bootstrap samples for use with
construction of confidence distributions, confidence curves,
P-values, etc.
}
\details{
bootcurve also estimates the bias-adjustment and acceleration
parameters for BCa bootstrap confidence interval.
}
